debug ffmpeg
